Feat: Create the initial InternalDataAgent (#2)
Creates an initial version of the agent using AutoGen's AssistantAgent, as well as an initial evaluation for this set up. The PR has been squashed as a series of commits alter the folder structure several times in a row, and I want a single commit with this finished set up.

Note that AssistantAgent is not intended for usage in the final system, as the class is intended for someone to prototype and get used to AutoGen as a library.